Bugs Fixed
Bugs Discussed:

Bug - 1: Android- On the update status page edit function is not working -- Not Implemented

Bug - 2: Vendor Signup & reg Business label remove and User name & Email & Phone No Data  Bind

Bug - 3: Booking Date not displaying -- Completed

Bug - 4: Track order edit remove once dispatched -- Completed

Bug - 5: Tracking id: not display once dispatched -- Completed

Bug - 6: Replace green tick mark by red cross --> Icon Need -- Completed ( Fixed with Temporary Icon)

Bug - 7: Cancelled Order Reason not show in track update changes  -- Completed

Bug - 8: Remove edit in track order Cancelled -- Completed

Bug - 9: Return order -- process flow requires clarification

+++ app/src/main/java/com/petfolio/infinitus/responsepojo/PetVendorOrderResponse.java
@@ -84,7 +84,7 @@ public class PetVendorOrderResponse {
         private String order_id;
         private String product_name;
         private int product_quantity;
-        private double product_price;
+        private int product_price;
         private String prodcut_image;
         private String date_of_booking;
         private String status;
@@ -173,11 +173,11 @@ public class PetVendorOrderResponse {
         }
 
 
-        public double getProduct_price() {
+        public int getProduct_price() {
             return product_price;
         }
 
-        public void setProduct_price(double product_price) {
+        public void setProduct_price(int product_price) {
             this.product_price = product_price;
 
         }
