+++ app/src/main/java/com/petfolio/infinitus/responsepojo/ShopDashboardResponse.java
@@ -134,7 +134,7 @@ public class ShopDashboardResponse {
             private String _id;
             private String product_img;
             private String product_title;
-            private double product_price;
+            private int product_price;
             private int product_discount;
             private boolean product_fav;
             private double product_rating;
@@ -171,11 +171,11 @@ public class ShopDashboardResponse {
             }
 
 
-            public double getProduct_price() {
+            public int getProduct_price() {
                 return product_price;
             }
 
-            public void setProduct_price(double product_price) {
+            public void setProduct_price(int product_price) {
                 this.product_price = product_price;
 
             }
@@ -285,7 +285,7 @@ public class ShopDashboardResponse {
                 private String _id;
                 private String product_img;
                 private String product_title;
-                private double product_price;
+                private int product_price;
                 private int product_discount;
                 private boolean product_fav;
                 private double product_rating;
@@ -322,11 +322,11 @@ public class ShopDashboardResponse {
                 }
 
 
-                public double getProduct_price() {
+                public int getProduct_price() {
                     return product_price;
                 }
 
-                public void setProduct_price(double product_price) {
+                public void setProduct_price(int product_price) {
                     this.product_price = product_price;
 
                 }
